+++ crates/board/src/platform/protocol.rs
@@ -1,0 +1,44 @@
+// Copyright 2024 Google LLC
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+//! Platform protocol interface.
+
+use alloc::boxed::Box;
+
+use crate::Error;
+
+/// Request received.
+#[derive(Debug, PartialEq, Eq)]
+pub struct Event;
+
+impl<B: crate::Api> From<Event> for crate::Event<B> {
+    fn from(event: Event) -> Self {
+        super::Event::Protocol(event).into()
+    }
+}
+
+/// Platform protocol interface.
+pub trait Api {
+    /// Reads the last request, if any.
+    fn read() -> Result<Option<Box<[u8]>>, Error>;
+
+    /// Writes a response for the last request.
+    fn write(response: &[u8]) -> Result<(), Error>;
+
+    /// Enables an event to be triggered when a request is received.
+    fn enable() -> Result<(), Error>;
+
+    /// Disables events from being triggered.
+    fn disable() -> Result<(), Error>;
+}