+++ crates/scheduler/src/event.rs
@@ -25,6 +25,8 @@ use crate::Scheduler;
 
 #[cfg(feature = "board-api-button")]
 pub mod button;
+#[cfg(feature = "internal-board-api-platform")]
+pub mod platform;
 #[cfg(feature = "internal-board-api-radio")]
 pub mod radio;
 #[cfg(feature = "board-api-timer")]
@@ -46,6 +48,8 @@ pub struct InstId;
 pub enum Key<B: Board> {
     #[cfg(feature = "board-api-button")]
     Button(button::Key<B>),
+    #[cfg(feature = "internal-board-api-platform")]
+    Platform(platform::Key),
     #[cfg(feature = "internal-board-api-radio")]
     Radio(radio::Key),
     #[cfg(feature = "board-api-timer")]
@@ -88,6 +92,14 @@ impl<'a, B: Board> From<&'a Event<B>> for Key<B> {
             #[cfg(feature = "board-api-button")]
             Event::Button(event) => {
                 or_unreachable!("applet-api-button", [event], Key::Button(event.into()))
+            }
+            #[cfg(feature = "internal-board-api-platform")]
+            Event::Platform(event) => {
+                or_unreachable!(
+                    "internal-applet-api-platform",
+                    [event],
+                    Key::Platform(event.into())
+                )
             }
             #[cfg(feature = "internal-board-api-radio")]
             Event::Radio(event) => {
@@ -150,6 +162,10 @@ pub fn process<B: Board>(scheduler: &mut Scheduler<B>, event: Event<B>) {
         #[cfg(feature = "board-api-button")]
         Event::Button(event) => {
             or_unreachable!("applet-api-button", [event], button::process(event, &mut params))
+        }
+        #[cfg(feature = "internal-board-api-platform")]
+        Event::Platform(event) => {
+            or_unreachable!("internal-applet-api-platform", [event], platform::process(event))
         }
         #[cfg(feature = "internal-board-api-radio")]
         Event::Radio(event) => {