+++ crates/board/CHANGELOG.md
@@ -9,6 +9,7 @@
 
 ### Minor
 
+- Add `platform::Api::Protocol` for the platform protocol and its `api-platform-protocol` feature
 - Add `crypto::{HashApi,HmacApi}` helpers for `crypto::{Hash,Hmac}`
 - Add `crypto::NoError` helper for `crypto::WithError`
 - Add `crypto::WithError` to add error support to RustCrypto (fix #176)