+++ crates/scheduler/CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Minor
 
+- Support `platform::protocol` with the `{applet,board}-api-platform-protocol` features
 - Change `gpio` module to never trap and return an error instead
 - Change `button::{register, unregister}()` to never trap and return an error instead
 - Migrate to `Id::new` returning `Result` instead of `Option`
@@ -129,4 +130,4 @@
 
 ## 0.1.0
 
-<!-- Increment to skip CHANGELOG.md test: 34 -->
+<!-- Increment to skip CHANGELOG.md test: 35 -->